refactor(backend): 共有カーネル shared モジュールを成形
横断的設施を旧フラットパッケージから com.kizuna.shared へ帰位:
- shared/tenancy: TenantContext、TenantIdInterceptor、TenantFilterEnable、TenantScoped
- shared/web: RequestCorrelationFilter
- shared/config: AppProperties、CacheConfig、RedisConfig、JacksonConfig
- shared/exception: ServiceException、CommonExceptionHandler
- shared/persistence: BaseEntity（central 系）、SnowflakeId、SnowflakeIdGenerator
- shared は OPEN モジュール（全モジュールから参照される共有カーネル）
- 空になった com.kizuna.exception を ModularityTests の除外リストから削除
- Jacoco 除外に shared/** を追加（移動前と同じ除外基準を維持、PR9 で見直し）

移動しないもの（依存が汚れているため各モジュール移行時に帰位）:
- JwtUtil / JwtAuthenticationFilter / SecurityConfig → auth モジュール（#195）
- MaintenanceModeInterceptor / WebMvcConfig → settings モジュール（#197）
- tenant 側 BaseEntity（Tenant への @ManyToOne が D3 の ID 参照化待ち）

Refs #193
