+++ backend/src/main/java/com/kizuna/model/entity/central/SystemConfig.java
@@ -1,5 +1,6 @@
 package com.kizuna.model.entity.central;
 
+import com.kizuna.shared.persistence.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
