+++ backend/src/main/java/com/kizuna/model/entity/central/security/CentralRole.java
@@ -1,6 +1,6 @@
 package com.kizuna.model.entity.central.security;
 
-import com.kizuna.model.entity.central.BaseEntity;
+import com.kizuna.shared.persistence.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
