+++ backend/src/main/java/com/kizuna/model/entity/central/tenant/Tenant.java
@@ -1,6 +1,6 @@
 package com.kizuna.model.entity.central.tenant;
 
-import com.kizuna.model.entity.central.BaseEntity;
+import com.kizuna.shared.persistence.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
