+++ backend/src/main/java/com/kizuna/service/CustomUserDetailsService.java
@@ -1,12 +1,12 @@
 package com.kizuna.service;
 
-import com.kizuna.config.interceptor.TenantContext;
 import com.kizuna.model.entity.central.security.CentralPermission;
 import com.kizuna.model.entity.central.security.CentralRole;
 import com.kizuna.model.entity.tenant.security.TenantPermission;
 import com.kizuna.model.entity.tenant.security.TenantRole;
 import com.kizuna.repository.central.CentralUserRepository;
 import com.kizuna.repository.tenant.TenantUserRepository;
+import com.kizuna.shared.tenancy.TenantContext;
 import io.jsonwebtoken.lang.Collections;
 import java.util.Collection;
 import java.util.List;
