+++ backend/src/main/java/com/kizuna/service/central/config/SystemConfigServiceImpl.java
@@ -1,11 +1,11 @@
 package com.kizuna.service.central.config;
 
-import com.kizuna.exception.ServiceException;
 import com.kizuna.mapper.central.SystemConfigMapper;
 import com.kizuna.model.dto.central.config.SystemConfigResponse;
 import com.kizuna.model.dto.central.config.SystemConfigUpdateRequest;
 import com.kizuna.model.entity.central.SystemConfig;
 import com.kizuna.repository.central.SystemConfigRepository;
+import com.kizuna.shared.exception.ServiceException;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
