+++ backend/src/main/java/com/kizuna/shared/config/AppProperties.java
@@ -1,4 +1,4 @@
-package com.kizuna.config;
+package com.kizuna.shared.config;
 
 import java.util.List;
 import lombok.Getter;
