+++ backend/src/main/java/com/kizuna/shared/package-info.java
@@ -1,0 +1,4 @@
+/** 共有カーネル: テナントコンテキスト、共通エンティティ基盤、横断的設定・例外。 全モジュールから参照されるため OPEN モジュールとして公開する。 */
+@org.springframework.modulith.ApplicationModule(
+    type = org.springframework.modulith.ApplicationModule.Type.OPEN)
+package com.kizuna.shared;
