+++ backend/src/main/java/com/kizuna/shared/persistence/SnowflakeIdGenerator.java
@@ -1,4 +1,4 @@
-package com.kizuna.model.entity.tenant;
+package com.kizuna.shared.persistence;
 
 import java.io.Serializable;
 import java.time.Instant;
