+++ backend/src/main/java/com/kizuna/shared/tenancy/TenantContext.java
@@ -1,4 +1,4 @@
-package com.kizuna.config.interceptor;
+package com.kizuna.shared.tenancy;
 
 import org.springframework.stereotype.Component;
