+++ backend/src/test/java/com/kizuna/controller/central/CentralConfigControllerTest.java
@@ -8,12 +8,12 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.kizuna.config.JacksonConfig;
-import com.kizuna.config.interceptor.TenantContext;
-import com.kizuna.exception.ServiceException;
 import com.kizuna.model.dto.central.config.SystemConfigResponse;
 import com.kizuna.model.dto.central.config.SystemConfigUpdateRequest;
 import com.kizuna.service.central.config.SystemConfigService;
+import com.kizuna.shared.config.JacksonConfig;
+import com.kizuna.shared.exception.ServiceException;
+import com.kizuna.shared.tenancy.TenantContext;
 import com.kizuna.utils.JwtUtil;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
