+++ backend/src/test/java/com/kizuna/service/central/config/SystemConfigServiceImplTest.java
@@ -8,12 +8,12 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.kizuna.exception.ServiceException;
 import com.kizuna.mapper.central.SystemConfigMapper;
 import com.kizuna.model.dto.central.config.SystemConfigResponse;
 import com.kizuna.model.dto.central.config.SystemConfigUpdateRequest;
 import com.kizuna.model.entity.central.SystemConfig;
 import com.kizuna.repository.central.SystemConfigRepository;
+import com.kizuna.shared.exception.ServiceException;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
